# core-api/legacy_scripts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Financial Analysis API",
...
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Production server (update with actual URL)"
    }
  ],
  "paths": {
...
    "/api/categories": {
      "post": {
        "tags": [
          "categories"
        ],
        "summary": "Create Category",
        "description": "Create a new category.\n\nArgs:\n    category: Category data\n    db: Database session\n    \nReturns:\n    CategoryResponse: Created category",
        "operationId": "create_category_api_categories_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "List Categories",
        "description": "List categories with optional filters.\n\nArgs:\n    category_type: Category type filter\n    parent_category_id: Parent category ID filter\n    include_transfer_categories: Include transfer categories\n    search_text: Search text\n    db: Database session\n    \nReturns:\n    CategoryListResponse: List of categories",
        "operationId": "list_categories_api_categories_get",
        "parameters": [
          {
            "name": "category_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(Income|Expense|Both)$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Category type filter",
              "title": "Category Type"
            },
            "description": "Category type filter"
          },
          {
            "name": "parent_category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Parent category ID filter (null for root categories)",
              "title": "Parent Category Id"
            },
            "description": "Parent category ID filter (null for root categories)"
          },
          {
            "name": "include_transfer_categories",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include transfer categories",
              "default": true,
              "title": "Include Transfer Categories"
            },
            "description": "Include transfer categories"
          },
          {
            "name": "search_text",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in category name or description",
              "title": "Search Text"
            },
            "description": "Search in category name or description"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/{category_id}": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "Get Category",
        "description": "Get category by ID.\n\nArgs:\n    category_id: Category ID\n    db: Database session\n    \nReturns:\n    CategoryResponse: Category data",
        "operationId": "get_category_api_categories__category_id__get",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "categories"
        ],
        "summary": "Update Category",
        "description": "Update category.\n\nArgs:\n    category_id: Category ID\n    category: Updated category data\n    db: Database session\n    \nReturns:\n    CategoryResponse: Updated category",
        "operationId": "update_category_api_categories__category_id__put",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "categories"
        ],
        "summary": "Delete Category",
        "description": "Delete category.\n\nArgs:\n    category_id: Category ID\n    force: Force delete\n    db: Database session\n    \nReturns:\n    SuccessResponse: Success message",
        "operationId": "delete_category_api_categories__category_id__delete",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Force delete even if category has transactions or children",
              "default": false,
              "title": "Force"
            },
            "description": "Force delete even if category has transactions or children"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/root": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "Get Root Categories",
        "description": "Get root categories (categories without parent).\n\nArgs:\n    category_type: Category type filter\n    db: Database session\n    \nReturns:\n    CategoryListResponse: List of root categories",
        "operationId": "get_root_categories_api_categories_root_get",
        "parameters": [
          {
            "name": "category_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(Income|Expense|Both)$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Category type filter",
              "title": "Category Type"
            },
            "description": "Category type filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/{category_id}/children": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "Get Child Categories",
        "description": "Get child categories of a parent category.\n\nArgs:\n    category_id: Parent category ID\n    db: Database session\n    \nReturns:\n    CategoryListResponse: List of child categories",
        "operationId": "get_child_categories_api_categories__category_id__children_get",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryCreate": {
        "properties": {
          "category_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Category Name",
            "description": "Category name"
          },
          "category_type": {
            "type": "string",
            "pattern": "^(Income|Expense|Both)$",
            "title": "Category Type",
            "description": "Category type"
          },
          "parent_category_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Category Id",
            "description": "Parent category ID"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Category description"
          },
          "is_transfer_category": {
            "type": "boolean",
            "title": "Is Transfer Category",
            "description": "Is transfer category",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "category_name",
          "category_type"
        ],
        "title": "CategoryCreate",
        "description": "Schema for creating a category."
      },
...
      "CategoryListResponse": {
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryResponse"
            },
            "type": "array",
            "title": "Categories",
            "description": "List of categories"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of categories"
          }
        },
        "type": "object",
        "required": [
          "categories",
          "total"
        ],
        "title": "CategoryListResponse",
        "description": "Schema for category list response."
      },
      "CategoryResponse": {
        "properties": {
          "category_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Category Name",
            "description": "Category name"
          },
          "category_type": {
            "type": "string",
            "pattern": "^(Income|Expense|Both)$",
            "title": "Category Type",
            "description": "Category type"
          },
          "parent_category_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Category Id",
            "description": "Parent category ID"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Category description"
          },
          "is_transfer_category": {
            "type": "boolean",
            "title": "Is Transfer Category",
            "description": "Is transfer category",
            "default": false
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id",
            "description": "Category ID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "category_name",
          "category_type",
          "category_id",
          "created_at"
        ],
        "title": "CategoryResponse",
        "description": "Schema for category response."
      },
...
      "CategoryUpdate": {
        "properties": {
          "category_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Name",
            "description": "Category name"
          },
          "category_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(Income|Expense|Both)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Type",
            "description": "Category type"
          },
          "parent_category_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Category Id",
            "description": "Parent category ID"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Category description"
          },
          "is_transfer_category": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Transfer Category",
            "description": "Is transfer category"
          }
        },
        "type": "object",
        "title": "CategoryUpdate",
        "description": "Schema for updating a category."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SuccessResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Success message"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Additional data"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "SuccessResponse",
        "description": "Success response schema."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication (to be implemented)"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication (to be implemented)"
      }
    },
...
  },
...
}